Separate content-indexing operations into separate component and implement direct content access component for content generators to use, which will use the indexing.
