+++ embedders/savant/src/main/resources/META-INF/beans.xml
@@ -22,7 +22,8 @@
   
   <decorators>
     <class>org.commonjava.indy.autoprox.data.AutoProxDataManagerDecorator</class>
-    <class>org.commonjava.indy.content.index.ContentIndexingContentManagerDecorator</class>
+    <class>org.commonjava.indy.content.index.IndexingContentManagerDecorator</class>
+    <class>org.commonjava.indy.content.index.IndexingDirectContentAccessDecorator</class>
   </decorators>
       
 </beans>
